modificando sueldo a double y agregando imagen

+++ Models/Formulario.cs
@@ -2,9 +2,9 @@ using System;
 namespace Tarea_progra.Models
 {
     public class Formulario{
-        public int sueldo{ get; set; }
+        public double sueldo{ get; set; }
 
-        public int resultado{ get; set; }
+        public double resultado{ get; set; }
         public int meses{ get; set; }
 
         public string mensaje{ get; set; }
